resolved comments and refactored code
