+++ crates/common_utils/src/types/user.rs
@@ -1,0 +1,8 @@
+/// Theme related types
+pub mod theme;
+
+/// User related types
+pub mod user;
+
+pub use theme::*;
+pub use user::*;
