+++ crates/router/src/utils/user.rs
@@ -12,10 +12,7 @@ use redis_interface::RedisConnectionPool;
 use router_env::env;
 
 use crate::{
-    consts::user::{
-        LINEAGE_CONTEXT_PREFIX, LINEAGE_CONTEXT_TIME_EXPIRY_IN_SECS, REDIS_SSO_PREFIX,
-        REDIS_SSO_TTL,
-    },
+    consts::user::{REDIS_SSO_PREFIX, REDIS_SSO_TTL},
     core::errors::{StorageError, UserErrors, UserResult},
     routes::SessionState,
     services::{
@@ -23,7 +20,7 @@ use crate::{
         authorization::roles::RoleInfo,
     },
     types::{
-        domain::{self, LineageContext, MerchantAccount, UserFromStorage},
+        domain::{self, MerchantAccount, UserFromStorage},
         transformers::ForeignFrom,
     },
 };
@@ -362,50 +359,4 @@ pub async fn validate_email_domain_auth_type_using_db(
             .any(|auth_method| auth_method.auth_type == required_auth_type))
     .then_some(())
     .ok_or(UserErrors::InvalidUserAuthMethodOperation.into())
-}
-
-pub async fn get_lineage_context_from_cache(
-    state: &SessionState,
-    user_id: &str,
-) -> UserResult<Option<LineageContext>> {
-    let connection = get_redis_connection_for_global_tenant(state)?;
-    let key = format!("{}{}", LINEAGE_CONTEXT_PREFIX, user_id);
-    let lineage_context = connection
-        .get_key::<Option<String>>(&key.into())
-        .await
-        .change_context(UserErrors::InternalServerError)
-        .attach_printable("Failed to get lineage context from redis")?;
-
-    match lineage_context {
-        Some(json_str) => {
-            let ctx = serde_json::from_str::<LineageContext>(&json_str)
-                .change_context(UserErrors::InternalServerError)
-                .attach_printable("Failed to deserialize LineageContext from JSON")?;
-            Ok(Some(ctx))
-        }
-        None => Ok(None),
-    }
-}
-
-pub async fn set_lineage_context_in_cache(
-    state: &SessionState,
-    user_id: &str,
-    lineage_context: LineageContext,
-) -> UserResult<()> {
-    let connection = get_redis_connection_for_global_tenant(state)?;
-    let key = format!("{}{}", LINEAGE_CONTEXT_PREFIX, user_id);
-    let serialized_lineage_context: String = serde_json::to_string(&lineage_context)
-        .change_context(UserErrors::InternalServerError)
-        .attach_printable("Failed to serialize LineageContext")?;
-    connection
-        .set_key_with_expiry(
-            &key.into(),
-            serialized_lineage_context,
-            LINEAGE_CONTEXT_TIME_EXPIRY_IN_SECS,
-        )
-        .await
-        .change_context(UserErrors::InternalServerError)
-        .attach_printable("Failed to set lineage context in redis")?;
-
-    Ok(())
 }
