+++ crates/router/src/utils/user/theme.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use common_enums::EntityType;
-use common_utils::{ext_traits::AsyncExt, id_type, types::theme::ThemeLineage};
+use common_utils::{ext_traits::AsyncExt, id_type, types::user::ThemeLineage};
 use diesel_models::user::theme::Theme;
 use error_stack::ResultExt;
 use hyperswitch_domain_models::merchant_key_store::MerchantKeyStore;
